// src/types/common/database.types.ts
export type Database = {
// ...
  core: {
    Tables: {
// ...
      profiles: {
        Row: {
          barangay: string | null
          birthdate: string | null
          contact_number: string | null
          created_at: string
          firstname: string | null
          gender: Database["core"]["Enums"]["gender_type"] | null
          geographic: string | null
          id: string
          is_4ps: boolean | null
          is_pwd: boolean | null
          last_login: string | null
          lastname: string | null
          middlename: string | null
          province: string | null
          region: string | null
          role: Database["core"]["Enums"]["user_role"] | null
          status: Database["core"]["Enums"]["status_type"] | null
          updated_at: string | null
          username: string | null
        }
        Insert: {
          barangay?: string | null
          birthdate?: string | null
          contact_number?: string | null
          created_at?: string
          firstname?: string | null
          gender?: Database["core"]["Enums"]["gender_type"] | null
          geographic?: string | null
          id?: string
          is_4ps?: boolean | null
          is_pwd?: boolean | null
          last_login?: string | null
          lastname?: string | null
          middlename?: string | null
          province?: string | null
          region?: string | null
          role?: Database["core"]["Enums"]["user_role"] | null
          status?: Database["core"]["Enums"]["status_type"] | null
          updated_at?: string | null
          username?: string | null
        }
        Update: {
          barangay?: string | null
          birthdate?: string | null
          contact_number?: string | null
          created_at?: string
          firstname?: string | null
          gender?: Database["core"]["Enums"]["gender_type"] | null
          geographic?: string | null
          id?: string
          is_4ps?: boolean | null
          is_pwd?: boolean | null
          last_login?: string | null
          lastname?: string | null
          middlename?: string | null
          province?: string | null
          region?: string | null
          role?: Database["core"]["Enums"]["user_role"] | null
          status?: Database["core"]["Enums"]["status_type"] | null
          updated_at?: string | null
          username?: string | null
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      gender_type:
        | "woman"
        | "man"
        | "cisgender_woman"
        | "cisgender_man"
        | "transgender_woman"
        | "transgender_man"
        | "non_binary"
        | "genderqueer"
        | "genderfluid"
        | "agender"
        | "two_spirit"
        | "intersex"
        | "different_identity"
        | "prefer_not_to_say"
      status_type: "pending" | "approved" | "rejected" | "active" | "inactive"
      user_role:
        | "admin"
        | "applicant"
        | "company_owner"
        | "company_member"
        | "provincial_peso"
        | "municipal_peso"
        | "dole"
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
// ...
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never